+++ book/src/applet/prelude/usb.rs
@@ -20,8 +20,7 @@
 
 //{ ANCHOR: all
 #![no_std]
-
-extern crate alloc;
+wasefire::applet!();
 
 use alloc::rc::Rc;
 use alloc::string::String;
@@ -29,10 +28,7 @@ use alloc::{format, vec};
 use core::cell::Cell;
 use core::time::Duration;
 
-use wasefire::*;
-
-#[no_mangle]
-pub extern "C" fn main() {
+fn main() {
     //{ ANCHOR: state
     let mut level = 3; // length of the string to remember
     let mut prompt = "Press ENTER when you are ready.";