+++ crates/prelude/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
+### Minor
+
+- Add applet!() macro
+
 ## 0.1.0